Move remaining unit tests
