+++ bundles/org.openhab.core.automation/pom.xml
@@ -29,6 +29,12 @@
       <artifactId>org.openhab.core.thing</artifactId>
       <version>${project.version}</version>
     </dependency>
+    <dependency>
+      <groupId>org.openhab.core.bundles</groupId>
+      <artifactId>org.openhab.core.test</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 
 </project>
