+++ bundles/org.openhab.core.automation/src/test/java/org/openhab/core/automation/internal/ConnectionValidatorTest.java
