+++ bundles/org.openhab.core.config.core/src/test/java/org/eclipse/smarthome/config/core/status/ConfigStatusServiceTest.java
