+++ bundles/org.openhab.core.config.discovery/src/test/java/org/eclipse/smarthome/config/discovery/internal/AutomaticInboxProcessorTest.java
