+++ bundles/org.openhab.core.io.rest.core/pom.xml
@@ -34,6 +34,12 @@
       <artifactId>org.openhab.core.io.rest</artifactId>
       <version>${project.version}</version>
     </dependency>
+    <dependency>
+      <groupId>org.openhab.core.bundles</groupId>
+      <artifactId>org.openhab.core.test</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 
 </project>
