+++ bundles/org.openhab.core.io.rest.core/src/test/java/org/eclipse/smarthome/io/rest/core/thing/EnrichedThingDTOMapperTest.java
