+++ itests/org.openhab.core.automation.tests/src/main/java/org/openhab/core/automation/internal/RuleRegistryTest.java
@@ -34,7 +34,7 @@ import org.openhab.core.automation.RuleRegistry;
  */
 public class RuleRegistryTest extends JavaOSGiTest {
 
-    RuleRegistry ruleRegistry;
+    private RuleRegistry ruleRegistry;
 
     @Before
     public void setup() {
